Add a new Rules module, fixes and tweaks.
Thr new Rules module will be used for /rules and /setrules in
larger community chats. Rules can be anything, with the
formatting preserved, and even support markdown.

Also tweaked welcome messages, adding a {chatname} option, and enabling
markdown support.

Finally, also avoid escaping characters in firstnames where markdown
characters are present BUT are already in a markdown URL - this was
happening in the /slaps command.